refactor[frontend]:useCategoriesの定義
