+++ apps/frontend/src/hooks/useCategories.ts
@@ -1,0 +1,22 @@
+import { useEffect, useState } from "react";
+import { fetchCategories, type Category } from "@/api";
+
+export default function useCategories() {
+  const [categories, setCategories] = useState<Category[]>([]);
+  const [selectedCategories, setSelectedCategories] = useState<Category[]>([]);
+
+  useEffect(() => {
+    const loadCategories = async () => {
+      try {
+        const data = await fetchCategories();
+        setCategories(data);
+        setSelectedCategories(data);
+      } catch (error) {
+        console.error("Failed to load categories:", error);
+      }
+    };
+    loadCategories();
+  }, []);
+
+  return { categories, selectedCategories, setSelectedCategories };
+}
